Extracted shared scoring engine into AbstractScoringTask base
* New OspreySharp/Tasks/AbstractScoringTask.cs holds the
  shared scoring engine: RunCoelutionScoring + ScoreWindow +
  ScoreCandidate + all feature-compute helpers (Compute*),
  library prep (LoadLibrary, GenerateDecoys, BuildDecoyFromSequence),
  ExtractIsolationWindows, ExtractTopNFragmentXics, dedup helpers
  (DeduplicateDoubleCounting, DeduplicatePairs), and the static
  utilities (BinarySearchLowerBound, FindNearestMs1, PearsonOverRange,
  HasTopNFragmentMatch, etc.) -- 33 methods total
* Static fields and constants travel with the methods:
  s_calXcorrScorer, s_mzmlReadGate, SG_WEIGHTS, _top6MzCache,
  TotalOrderComparer, NUM_PIN_FEATURES, BASE_ID_MASK,
  CAL_TOP_N_FRAGMENTS, DIAG_PEPTIDE, plus the WindowTiming
  nested class
* PerFileScoringTask, PerFileRescoreTask, and FirstJoinTask all
  inherit AbstractScoringTask; no more _pipeline back-references
  for engine calls
* AnalysisPipeline also inherits AbstractScoringTask as a pragmatic
  shortcut so AnalysisPipeline.Stage6Rescore.cs (still on the
  AnalysisPipeline partial) can call inherited engine methods.
  AnalysisPipeline.Run(PipelineContext) and Name overrides throw
  NotSupportedException -- AnalysisPipeline is the orchestrator,
  not a runnable task
* PerFileRescoreTask plants its _ctx onto the AnalysisPipeline
  back-reference before invoking ExecuteStage6Rescore so the
  inherited engine methods see a non-null _ctx through `this`.
  RunWorker (the worker-mode CLI entry) initializes its own _ctx
  on entry from config + the static log sinks
* AnalysisPipeline.cs shrinks from 3131 lines to 257 lines (96%
  reduction from original 7054). Holds Run + log sinks +
  FormatDuration + F10
* Stellar 3-file snapshot regression PASS at every stage

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
